Modify the prefix adding function for colliding samples
